refactor: extract NovelReader logic into useNovelReader hook
- Created `useNovelReader.ts` to encapsulate the EPUB loading, theme application, navigation handling, and DOM element management previously located in `NovelReader.tsx`.
- Updated `NovelReader.tsx` to utilize the new `useNovelReader` hook, simplifying the component to primarily handle the layout and the `viewerRef`.
- Improved state handling in the hook by using a `contextRef` to ensure async operations correctly access the latest application state and callbacks without requiring constant re-initialization of the book viewer.
